fix: resolve 6 additional bugs found by review agents
CRITICAL fixes:
- utils/helpers.go: SplitMessage infinite loop with multi-byte chars at position 0
  - Use utf8.RuneStart() and utf8.DecodeRuneInString() for proper UTF-8 handling
  - When cutAt reaches 0, find end of first complete character instead of looping
- retry/retry.go: computeDelay overflow on 32-bit platforms
  - Use int64(1)<<exp instead of 1<<exp to avoid undefined behavior
- ai/client.go: CreateVisionCompletionWithTools missing tool call stripping
  - Add same unprocessed tool call stripping as CreateChatCompletionWithTools

MEDIUM fixes:
- memory/consolidation.go: analyzeConversationBatch now uses extractJSONFromLLMResponse + sanitizeJSON
  - Consistent with analyzeConversation, handles Go 1.25 JSON compatibility
- web/handlers_config.go: Save failure now rolls back atomic store
  - Prevents in-memory config from diverging from YAML on disk
- ai/vision/vision.go: Temperature=0 now treated as valid value
  - Changed <= 0 to < 0 to allow deterministic output (temperature=0)

All tests pass.

+++ internal/retry/retry.go
@@ -128,11 +128,12 @@ func Retry[T any](ctx context.Context, maxRetries int, fn func(context.Context) 
 // delay = min(2^attempt * baseDelay, maxDelay)
 func computeDelay(attempt int, baseDelay, maxDelay time.Duration) time.Duration {
 	// Cap attempt to prevent integer overflow in bit shift
+	// Use int64(1) to avoid undefined behavior on 32-bit platforms where int is 32 bits
 	exp := uint(attempt)
 	if exp > 62 {
 		exp = 62
 	}
-	delay := min(time.Duration(1<<exp)*baseDelay, maxDelay)
+	delay := min(time.Duration(int64(1)<<exp)*baseDelay, maxDelay)
 	jitter := time.Duration(float64(delay) * 0.25 * (2.0*rand.Float64() - 1.0))
 	return delay + jitter
 }
